Add missing `return false` to fix incorrect log message
Because we didn't have a `return false` after detecting a non-match in
`checkRegexp`, we were logging a message claiming that the regexp was
empty and not being tested, when in fact it was non-empty but didn't
match. This commit adds the `return` statement to fix this. Note that
this doesn't change the functionality of the extension; it's just to
clean up an incorrect debug log message.