Do not default to block if neither regexp matches
Do not default to Block if neither regexp matches; only change the
content policy for a message if one or the other of the regexps
matches.

This reverts an unwanted change in behavior that was introduced in the
port to TB115.

If the user wants to default to blocking, then can set the block
regexp to "`.*`".