Restore previous behavior: only check policy for new messages
The desired behavior of the add-on is to check new messages as they
are received to decide whether to change their policy, and not to
update the policy on messages which already have a policy. This
behavior was consciously, intentionally chosen by the creator of the
add-on. It has three benefits:

* It minimizes the likelihood of conflict between this add-on and
  others that mess with the content policies on messages.
* It maximizes performance because messages only need to be checked
  once, when they are received.
* It prevents the content policy banner from flashing briefly when the
  message is displayed, because the policy is updated before the
  message is displayed.

The initial attempt to port this add-on to TB115 changed the behavior
to check all messages when they are displayed rather than just
checking newly received messages. This commit reverts to the previous,
intended behavior.

With this updated behavior, it does note makes sense to have a
preference indicating that remote content should only be enabled
temporarily on messages, so this preference has been removed.