Allow user to specify folders to be scanned on startup
Sometimes on startup Thunderbird starts fetching new messages from the
mail server before all add-ons have finished initializing. New
messages may therefore be fetched from the server before Remote
Content By Folder (RCBF) has started listening for new messages,
resulting in the the policies specified by the RCBF not being applied
to those messages.

To address this, the add-on now allowed the user to specify a regexp
indicating which folders should be scanned for new messages on
startup. This regexp is not set by default, since I don't want the
add-on to do anything that the user hasn't explicitly configured, but
the value a user is most likely to want to put in this setting is
`^Inbox$`, so all new messages in the inbox will have the proper RCBF
policy applied.

Note that a folder's name still has to match the allow or block regexp
for this new setting to have any effect on it.