Fix 2 false assumptions: message IDs come in order and we see all
For whatever reason, we apparently don't see all message IDs that the
backend allocates, and they also don't come to us in order, so we
can't just keep a maximum seen message ID and rely on that to
determine which messages we've already processed. So now, instead,
while we're scanning we keep a hash telling us which messages we've
seen, and then at the end of a scan that doesn't see any new messages,
we reset the baseline message ID and clear the hash so we don't waste
memory.