Big rewrite for better performance and resiliency
The "seen baseline" functionality wasn't working properly and
therefore we were scanning all of the messages in all the scanned
folders over and over every minute. I have now replaced it with a
sequence set implementation which keeps track of all the message IDs
we've already seen and doesn't rescan any of them.

There is new event log functionality which keeps up to 48 hours worth
of event logs internally, separate from the Thunderbird console, to
assist with troubleshooting in general. Specifically, I hope this will
help me confirm whether Thunderbird is generating onNewMailReceived
events properly for all new messages, and if not, give me enough data
to report something useful about the problem to the maintainers of
Thunderbird.