Don't anomalize a new mail notification coming in shortly after a scan
When we discover a message in a scan, and then shortly after we get
notified about the new message, we don't need to treat that as an
anomaly. It's an expected race condition when we happen to run a scan
while TB is in the process of generating the new mail notification.