Don't allow multiple checkSeenRecently to run concurrently
If there are so many messages seen recently that it takes more than 5
seconds to for `checkSeenRecently` to run then we could end up with
two running concurrently reporting the same issues, and if that puts
enough load on TB that it slows them down, each of the running
instances could get slowed down, compounding the problem and causing
it to spiral out of control and cause many checks to run in parallel
reporting the same issues.

Therefore we need to ensure that only one `checkSeenRecently` runs at
a time.